Finished Phonebook App

+++ Models/Phone_Numbers.cs
@@ -5,7 +5,7 @@ namespace Models;
 public class PhoneNumber
 {
     public int PhoneNumberId { get; set; }
-    public int TenDigitNumber { get; set; }
+    public long TenDigitNumber { get; set; }
     public string Location { get; set; }
 
     public int ContactId { get; set; }
